common: Move grid logic to browser-sdk
