+++ packages/browser-sdk/src/lib/react/Grid/layout/helpers.ts
