+++ packages/browser-sdk/src/lib/react/Grid/layout/types.ts
@@ -1,3 +1,5 @@
+import { ClientView } from "@whereby.com/core";
+
 export type Box = {
     top: number;
     left: number;
@@ -18,18 +20,6 @@ export type Bounds = {
 export type Frame = {
     origin: Origin;
     bounds: Bounds;
-};
-
-export type ClientView = {
-    id: string;
-    clientId: string;
-    displayName: string;
-    hasActivePresentation?: boolean;
-    stream?: MediaStream | null;
-    isLocalClient?: boolean;
-    isPresentation?: boolean;
-    isVideoEnabled?: boolean;
-    isAudioEnabled?: boolean;
 };
 
 export type CellView = {
