+++ packages/browser-sdk/src/lib/react/Grid/selector.ts
@@ -1,39 +1,74 @@
 import { createSelector } from "@reduxjs/toolkit";
 import {
-    selectCellViewsVideoGrid,
-    RemoteParticipant,
-    CellView,
-    StageLayout,
-    selectLayoutVideoStage,
-    selectCellViewsPresentationGrid,
-    selectCellViewsSubgrid,
-    Frame,
-    selectLayoutContainerFrame,
+    selectClientViewsInSubgrid,
+    selectLayoutClientAspectRatios,
+    selectClientViewsInPresentationGrid,
+    selectClientViewsInGrid,
 } from "@whereby.com/core";
+import { makeVideoCellView } from "./layout/cellView";
+import { type CellView } from "./layout/types";
 
-export type RemoteParticipantState = Omit<RemoteParticipant, "newJoiner" | "streams">;
+const selectCellViewsSubgrid = createSelector(
+    selectClientViewsInSubgrid,
+    selectLayoutClientAspectRatios,
+    (clientViews, clientAspectRatios) => {
+        return clientViews.map((client) => {
+            return makeVideoCellView({
+                client,
+                aspectRatio: clientAspectRatios[client.id],
+                avatarSize: 0,
+                cellPaddings: 0,
+                isSubgrid: true,
+            });
+        });
+    },
+);
+
+const selectCellViewsPresentationGrid = createSelector(
+    selectClientViewsInPresentationGrid,
+    selectLayoutClientAspectRatios,
+    (clientViews, clientAspectRatios) => {
+        return clientViews.map((client) => {
+            return makeVideoCellView({
+                client,
+                aspectRatio: clientAspectRatios[client.id],
+                avatarSize: 0,
+                cellPaddings: 0,
+            });
+        });
+    },
+);
+
+const selectCellViewsVideoGrid = createSelector(
+    selectLayoutClientAspectRatios,
+    selectClientViewsInGrid,
+    (clientAspectRatios, clientViews) => [
+        ...clientViews.map((client) => {
+            return makeVideoCellView({
+                client,
+                aspectRatio: clientAspectRatios[client.id],
+                avatarSize: 0,
+                cellPaddings: 10,
+            });
+        }),
+    ],
+);
 
 export interface VideoGridState {
     cellViewsVideoGrid: CellView[];
     cellViewsInPresentationGrid: CellView[];
     cellViewsInSubgrid: CellView[];
-    videoStage: StageLayout | null;
-    containerFrame: Frame;
 }
 
 export const selectVideoGridState = createSelector(
     selectCellViewsVideoGrid,
     selectCellViewsPresentationGrid,
     selectCellViewsSubgrid,
-    selectLayoutVideoStage,
-    selectLayoutContainerFrame,
-    (cellViewsVideoGrid, cellViewsInPresentationGrid, cellViewsInSubgrid, videoStage, containerFrame) => {
+    (cellViewsVideoGrid, cellViewsInPresentationGrid, cellViewsInSubgrid) => {
         const state: VideoGridState = {
             cellViewsVideoGrid,
             cellViewsInPresentationGrid,
             cellViewsInSubgrid,
-            videoStage,
-            containerFrame,
         };
 
         return state;
