+++ packages/core/src/utils/index.ts
@@ -3,7 +3,3 @@ export { default as fakeWebcamFrame } from "./fakeWebcamFrame";
 export { default as debounce } from "./debounce";
 export { default as parseRoomUrlAndSubdomain } from "./roomUrl";
 export { default as parseUnverifiedRoomKeyData } from "./roomKey";
-export * from "./layout/types";
-export * from "./layout/gridUtils";
-export * from "./layout/cellView";
-export * from "./layout/helpers";
